+++ python/paddle/fluid/tests/unittests/test_dynrnn_gradient_check.py
@@ -19,7 +19,7 @@ import random
 import collections
 import paddle.fluid as fluid
 import unittest
-from decorators import *
+from decorator_helper import *
 
 
 class Memory(object):
